File sending seems to be working.. Crude drawing preview enabled

+++ src/tgfx/ResponseHeader.java
@@ -15,7 +15,7 @@ public class ResponseHeader {
     public static String statusMessage = "";
     private static int statusCode = 0;
     private static int bufferAvailable = 0;
-    private static double lineNumber;
+    private static long lineNumber;
     
     public ResponseHeader(){
        
@@ -24,7 +24,7 @@ public class ResponseHeader {
     public  int getBufferAvailable() {
         return bufferAvailable;
     }
-    public  double getLineNumber() {
+    public  long getLineNumber() {
         return lineNumber;
     }
     public static int getStatusCode() {
@@ -38,27 +38,6 @@ public class ResponseHeader {
         statusMessage = responseNodeObject.getNode("r").getNode("sm").getText();
         statusCode = Integer.valueOf(responseNodeObject.getNode("r").getNode("sc").getText());
         bufferAvailable = Integer.valueOf(responseNodeObject.getNode("r").getNode("buf").getText());
-        lineNumber = Double.valueOf(responseNodeObject.getNode("r").getNode("ln").getText());
-    }
-    
+        lineNumber = Long.valueOf(responseNodeObject.getNode("r").getNode("ln").getText());
+    }    
 }
-
-
-//
-//   private String[] getStatusMessage(JsonRootNode json) {
-//        /**
-//         * This function parses all return status codes and messages before
-//         * anything else
-//         */
-//        String statusMessage;
-//        String statusCode;
-//        try {
-//            statusMessage = json.getNode("r").getNode("sm").getText();
-//            statusCode = json.getNode("r").getNode("sc").getText();
-//            String[] ret = {statusMessage, statusCode};
-//            return (ret);
-//        } catch (Exception ex) {
-//            String[] ret = {"JSON Invalid", "-1"};
-//            return (ret);
-//        }
-//    }
